fix(backend): fix FK violations and race conditions in document tests
Add INSERT INTO collections before process_upload/process_zip_upload calls
that were missing parent rows, causing documents_collection_id_fkey violations.
Add serial_test attribute to all document tests to prevent TRUNCATE race
conditions under parallel execution. Fix INT4/i64 type mismatch for
chunks.index column. Replace .ok() with .expect() on test setup queries.

Fix i64 → i32 type mismatch in save_chunk, get_chunks, and
get_active_chunks for the chunks.index column (PostgreSQL INTEGER).

Fix raw-string SQL escape in test_reload_document query.

+++ backend/src/modules/documents/repository.rs
@@ -290,7 +290,7 @@ impl DocumentRepository {
         sqlx::query(r#"INSERT INTO chunks (id, document_id, "index", text, is_active) VALUES ($1, $2, $3, $4, $5)"#)
             .bind(chunk.id)
             .bind(chunk.document_id)
-            .bind(chunk.index as i64)
+            .bind(chunk.index as i32)
             .bind(&chunk.text)
             .bind(chunk.is_active)
             .execute(&self.db)
@@ -302,7 +302,7 @@ impl DocumentRepository {
 
     /// Retrieve chunks by document ID, ordered by index.
     pub async fn get_chunks(&self, document_id: Uuid) -> Result<Vec<Chunk>, AppError> {
-        let rows = sqlx::query_as::<_, (uuid::Uuid, uuid::Uuid, i64, String, bool)>(
+        let rows = sqlx::query_as::<_, (uuid::Uuid, uuid::Uuid, i32, String, bool)>(
             r#"SELECT id, document_id, "index", text, is_active FROM chunks WHERE document_id = $1 ORDER BY "index""#
         )
         .bind(document_id)
@@ -429,7 +429,7 @@ impl DocumentRepository {
     pub async fn get_active_chunks(&self, document_id: Uuid) -> Result<Vec<Chunk>, AppError> {
         tracing::debug!("Fetching active chunks for document: {document_id}");
 
-        let rows = sqlx::query_as::<_, (uuid::Uuid, uuid::Uuid, i64, String, bool)>(
+        let rows = sqlx::query_as::<_, (uuid::Uuid, uuid::Uuid, i32, String, bool)>(
             r#"SELECT id, document_id, "index", text, is_active FROM chunks
                WHERE document_id = $1 AND is_active = TRUE
                ORDER BY "index""#,
